move complexTypeSolving
